BE/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
asyncpg==0.30.0
pydantic[email]==2.11.3
pydantic-settings==2.9.1
watchfiles==0.24.0
pyjwt==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.0
